Refactor chunks-to-matrix code and its deps
